basic react router refactor
